Step 1: widen brk / mmap_next / ForkSnapshot from u32 to u64
Internal storage and the corresponding setter/getter signatures move
from u32 to u64. User-space addresses still live below 4 GiB and the
syscall ABIs (sys_sbrk i32->u32, sys_mmap u32->u32) are unchanged —
the cast at the syscall boundary is lossless as long as user space
stays in the lower half. This is the type-foundation for the upcoming
upper-half user-VA layout migration.

  kernel/src/task/thread.rs
    - Thread.brk: u32 -> u64
    - Thread.mmap_next: u32 -> u64

  kernel/src/task/scheduler/fork.rs
    - ForkSnapshot.brk / .mmap_next: u32 -> u64
    - set_thread_mmap_next(val: u64)
    - exec_update_thread(... brk: u64 ...)
    - exec_update_thread mmap reset literal -> 0x7000_0000u64

  kernel/src/task/scheduler/thread_config.rs
    - set_thread_user_info(brk: u64)
    - current_thread_brk() -> u64
    - set_current_thread_brk(brk: u64)
    - current_thread_mmap_next() -> u64
    - set_current_thread_mmap_next(val: u64)

  kernel/src/memory/vma.rs
    - init_process(mmap_hint: u64)

  kernel/src/task/loader.rs
    - drop "result.brk as u32" / "brk as u32" casts
    - mmap_start computed as u64

  kernel/src/syscall/handlers/process.rs
    - sys_sbrk casts the u64 brk to u32 once at the boundary; rest
      of the function stays u32 (no behavior change)
    - sys_mmap drops the now-unnecessary "(base + aligned_size) as u32"
      cast on set_current_thread_mmap_next

  kernel/src/task/scheduler/debug_trace.rs
    - put_u32 still emits the low 32 bits of brk/mmap_next for ABI
      compatibility (taskmanager etc.)

Verified: build green, image boots in QEMU/KVM, AMID/CONFD/compositor
spawn cleanly, no exceptions.

+++ kernel/src/memory/vma.rs
@@ -59,7 +59,7 @@ static VMA_REGISTRY: Spinlock<Vec<ProcessVmas>> = Spinlock::new(Vec::new());
 ///
 /// Called from `load_and_run_with_args()` after the page directory is created.
 /// `mmap_hint` includes ASLR randomization.
-pub fn init_process(pd: PhysAddr, mmap_hint: u32) {
+pub fn init_process(pd: PhysAddr, mmap_hint: u64) {
     let mut reg = VMA_REGISTRY.lock();
     // Guard: don't double-init (shouldn't happen, but be safe).
     if reg.iter().any(|p| p.pd == pd) {
@@ -68,7 +68,7 @@ pub fn init_process(pd: PhysAddr, mmap_hint: u32) {
     reg.push(ProcessVmas {
         pd,
         vmas: BTreeMap::new(),
-        mmap_hint: (mmap_hint as u64).max(MMAP_BASE),
+        mmap_hint: mmap_hint.max(MMAP_BASE),
         mmap64_hint: MMAP64_BASE,
     });
 }
